+++ packages/provider/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ai-sdk/provider
 
+## 4.0.0-beta.0
+
+### Major Changes
+
+- 8359612: Start v7 pre-release
+
 ## 3.0.8
 
 ### Patch Changes